disaster_predictor/requirements.txt
pandas-gbq
pyarrow
db-dtypes
google-auth
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
earthengine-api
python-dotenv
urllib3<2
openmeteo-requests
timezonefinder
h3>=4.4.2
requests-cache
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.15.0
ollama
scipy
joblib
scikit-learn